requirements.txt
flask
requests
reportlab
python-docx
Flask-Session
